resource management moved to namespace and folder ResourceManagement
