+++ src/Redwood.Framework/Redwood.Framework.csproj
@@ -83,18 +83,18 @@
     <Compile Include="Binding\ValueBindingExpression.cs" />
     <Compile Include="Binding\ViewModelJTokenEvaluationVisitor.cs" />
     <Compile Include="Binding\ViewModelJTokenEvaluator.cs" />
-    <Compile Include="Configuration\InlineScriptResource.cs" />
-    <Compile Include="Configuration\IRedwoodResourceRepository.cs" />
-    <Compile Include="Configuration\RedwoodConfiguration.cs" />
+    <Compile Include="ResourceManagement\InlineScriptResource.cs" />
+    <Compile Include="ResourceManagement\IRedwoodResourceRepository.cs" />
+    <Compile Include="ResourceManagement\RedwoodConfiguration.cs" />
     <Compile Include="Configuration\RedwoodControlConfiguration.cs" />
     <Compile Include="Configuration\RedwoodMarkupConfiguration.cs" />
-    <Compile Include="Configuration\RedwoodResourceRepository.cs" />
+    <Compile Include="ResourceManagement\RedwoodResourceRepository.cs" />
     <Compile Include="Configuration\RedwoodSecurityConfiguration.cs" />
-    <Compile Include="Configuration\ResourceBase.cs" />
-    <Compile Include="Configuration\ResourceConfigurationCollectionNameAttribute.cs" />
-    <Compile Include="Configuration\ResourceRepositoryJsonConverter.cs" />
-    <Compile Include="Configuration\ScriptResource.cs" />
-    <Compile Include="Configuration\StylesheetResource.cs" />
+    <Compile Include="ResourceManagement\ResourceBase.cs" />
+    <Compile Include="ResourceManagement\ResourceConfigurationCollectionNameAttribute.cs" />
+    <Compile Include="ResourceManagement\ResourceRepositoryJsonConverter.cs" />
+    <Compile Include="ResourceManagement\ScriptResource.cs" />
+    <Compile Include="ResourceManagement\StylesheetResource.cs" />
     <Compile Include="Controls\Bootstrap\FormGroup.cs" />
     <Compile Include="Controls\Bootstrap\GlyphIcon.cs" />
     <Compile Include="Controls\Bootstrap\HorizontalForm.cs" />
