+++ src/Redwood.Framework/ResourceManagement/IRedwoodResourceRepository.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Redwood.Framework.Configuration
+namespace Redwood.Framework.ResourceManagement
 {
     public interface IRedwoodResourceRepository
     {
