+++ src/Redwood.Framework/ResourceManagement/ResourceConfigurationCollectionNameAttribute.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Redwood.Framework.Configuration
+namespace Redwood.Framework.ResourceManagement
 {
     [AttributeUsage(AttributeTargets.Class, Inherited = false, AllowMultiple = false)]
     public sealed class ResourceConfigurationCollectionNameAttribute : Attribute
